examples: добавлены примеры для IntroSat.Pico (#123)
* examples: добавлены примеры для IntroSat.Pico

* fix: исправление опечатки

* examples: исправление опечаток

* examples: добавлен пример для Pico CC1101/ReceiveSimple

* examples: исправлена опечатка в примере

* examples: исправлена опечатка в примере

* examples: исправлена опечатка в примере

* examples: исправлена ошибка в примере Pico/CC1101/ReceiveSimple

* examples: исправлен тип данных в примерах Pico/CC1101

* fix: исправлен тип аргумента CC1101::receive() size_t -> uint16_t

* fix: исправлен тип аргумента в примере Pico/CC1101/ReceiveSimple

* fix: исправлены опечатки в примерах ИК (#134)

---------

+++ src/Device/CC1101/CC1101.cpp
@@ -488,7 +488,7 @@ CC1101::Status CC1101::transmit(uint8_t *data, size_t length, uint8_t addr) {
   return ret;
 }
 
-CC1101::Status CC1101::receive(uint8_t *data, size_t length, size_t *read, uint8_t addr) {
+CC1101::Status CC1101::receive(uint8_t *data, size_t length, uint16_t *read, uint8_t addr) {
   if (length > 255) {
     return STATUS_LENGTH_TOO_BIG;
   }
