+++ src/Device/CC1101/CC1101.h
@@ -230,7 +230,7 @@ public:
 	void setSyncWord(uint16_t sync);
 
 	Status transmit(uint8_t *data, size_t length, uint8_t addr = 0);
-	Status receive(uint8_t *data, size_t length, size_t *read = nullptr,
+	Status receive(uint8_t *data, size_t length, uint16_t *read = nullptr,
 			uint8_t addr = 0);
 	int8_t getRSSI();
 	uint8_t getLQI();
